Create AuraEnabled_Annotation
There are only two specific things that make this method available to your Lightning Components code.

The @AuraEnabled annotation before the method declaration.
“Aura” is the name of the open source framework at the core of Lightning Components. 
You’ve seen it used in the namespace for some of the core tags, such as <aura:component>. 
Now you know where it comes from.

The static keyword. All @AuraEnabled methods must be static methods, and either public or global scope.
If these requirements remind you of remote methods for Visualforce’s JavaScript remoting feature, 
that’s not a coincidence. The requirements are the same, 
because the architecture is very similar at key points.